+++ packages/analytics-browser/src/browser-client.ts
@@ -35,6 +35,9 @@ import {
   isElementInteractionsEnabled,
   isPageViewTrackingEnabled,
   isNetworkTrackingEnabled,
+  isWebVitalsEnabled,
+  isFrustrationInteractionsEnabled,
+  getFrustrationInteractionsConfig,
 } from './default-tracking';
 import { convertProxyObjectToRealObject, isInstanceProxy } from './utils/snippet-helper';
 import { Context } from './plugins/context';
@@ -46,8 +49,9 @@ import { DEFAULT_SESSION_END_EVENT, DEFAULT_SESSION_START_EVENT } from './consta
 import { detNotify } from './det-notification';
 import { networkConnectivityCheckerPlugin } from './plugins/network-connectivity-checker';
 import { createBrowserJoinedConfigGenerator } from './config/joined-config';
-import { autocapturePlugin } from '@amplitude/plugin-autocapture-browser';
+import { autocapturePlugin, frustrationPlugin } from '@amplitude/plugin-autocapture-browser';
 import { plugin as networkCapturePlugin } from '@amplitude/plugin-network-capture-browser';
+import { webVitalsPlugin } from '@amplitude/plugin-web-vitals-browser';
 import { WebAttribution } from './attribution/web-attribution';
 
 /**
@@ -159,9 +163,19 @@ export class AmplitudeBrowser extends AmplitudeCore implements BrowserClient {
       await this.add(autocapturePlugin(getElementInteractionsConfig(this.config))).promise;
     }
 
+    if (isFrustrationInteractionsEnabled(this.config.autocapture)) {
+      this.config.loggerProvider.debug('Adding frustration interactions plugin');
+      await this.add(frustrationPlugin(getFrustrationInteractionsConfig(this.config))).promise;
+    }
+
     if (isNetworkTrackingEnabled(this.config.autocapture)) {
       this.config.loggerProvider.debug('Adding network tracking plugin');
       await this.add(networkCapturePlugin(getNetworkTrackingConfig(this.config))).promise;
+    }
+
+    if (isWebVitalsEnabled(this.config.autocapture)) {
+      this.config.loggerProvider.debug('Adding web vitals plugin');
+      await this.add(webVitalsPlugin()).promise;
     }
 
     this.initializing = false;