+++ packages/analytics-browser/src/config/joined-config.ts
@@ -100,6 +100,8 @@ export class BrowserJoinedConfigGenerator {
               pageViews: this.config.autocapture,
               sessions: this.config.autocapture,
               elementInteractions: this.config.autocapture,
+              webVitals: this.config.autocapture,
+              frustrationInteractions: this.config.autocapture,
               ...transformedAutocaptureRemoteConfig,
             };
           }