+++ packages/analytics-browser/test/config/joined-config.test.ts
@@ -104,6 +104,8 @@ describe('joined-config', () => {
           pageViews: true,
           attribution: true,
           elementInteractions: false,
+          webVitals: true,
+          frustrationInteractions: true,
         };
 
         await generator.initialize();
@@ -128,6 +130,7 @@ describe('joined-config', () => {
             pageViews: false,
             attribution: false,
             sessions: false,
+            webVitals: false,
           },
         };
         mockRemoteConfigFetch = {
@@ -146,6 +149,8 @@ describe('joined-config', () => {
           attribution: false,
           sessions: false,
           elementInteractions: true,
+          webVitals: false,
+          frustrationInteractions: true,
         };
 
         await generator.initialize();
@@ -163,6 +168,8 @@ describe('joined-config', () => {
           formInteractions: false,
           attribution: false,
           pageViews: false,
+          webVitals: false,
+          frustrationInteractions: false,
         },
         false,
       ])('should only enable elementInteractions if remote config only sets it to true', async (localAutocapture) => {
@@ -193,6 +200,8 @@ describe('joined-config', () => {
           attribution: false,
           pageViews: false,
           elementInteractions: true,
+          webVitals: false,
+          frustrationInteractions: false,
         };
 
         await generator.initialize();