+++ packages/session-replay-browser/src/config/local-config.ts
@@ -1,5 +1,10 @@
 import { Config, Logger, FetchTransport, LogLevel } from '@amplitude/analytics-core';
-import { DEFAULT_PERFORMANCE_CONFIG, DEFAULT_SAMPLE_RATE, DEFAULT_SERVER_ZONE } from '../constants';
+import {
+  DEFAULT_PERFORMANCE_CONFIG,
+  DEFAULT_SAMPLE_RATE,
+  DEFAULT_SERVER_ZONE,
+  DEFAULT_URL_CHANGE_POLLING_INTERVAL,
+} from '../constants';
 import { SessionReplayOptions, StoreType } from '../typings/session-replay';
 import {
   SessionReplayLocalConfig as ISessionReplayLocalConfig,
@@ -36,6 +41,9 @@ export class SessionReplayLocalConfig extends Config implements ISessionReplayLo
     script?: boolean;
     comment?: boolean;
   };
+  enableUrlChangePolling?: boolean;
+  urlChangePollingInterval?: number;
+  captureDocumentTitle?: boolean;
 
   constructor(apiKey: string, options: SessionReplayOptions) {
     const defaultConfig = getDefaultConfig();
@@ -60,6 +68,9 @@ export class SessionReplayLocalConfig extends Config implements ISessionReplayLo
     this.performanceConfig = options.performanceConfig || DEFAULT_PERFORMANCE_CONFIG;
     this.storeType = options.storeType ?? 'idb';
     this.applyBackgroundColorToBlockedElements = options.applyBackgroundColorToBlockedElements ?? false;
+    this.enableUrlChangePolling = options.enableUrlChangePolling ?? false;
+    this.urlChangePollingInterval = options.urlChangePollingInterval ?? DEFAULT_URL_CHANGE_POLLING_INTERVAL;
+    this.captureDocumentTitle = options.captureDocumentTitle ?? false;
 
     if (options.privacyConfig) {
       this.privacyConfig = options.privacyConfig;