+++ packages/session-replay-browser/CHANGELOG.md
@@ -3,6 +3,36 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.27.0](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/session-replay-browser@1.26.2...@amplitude/session-replay-browser@1.27.0) (2025-07-17)
+
+
+### Features
+
+* detect URL changes ([#1192](https://github.com/amplitude/Amplitude-TypeScript/issues/1192)) ([4957c16](https://github.com/amplitude/Amplitude-TypeScript/commit/4957c16df1f1b102cb2420d72b8f60847f49424a))
+
+
+
+
+
+## [1.26.2](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/session-replay-browser@1.26.1...@amplitude/session-replay-browser@1.26.2) (2025-07-15)
+
+**Note:** Version bump only for package @amplitude/session-replay-browser
+
+
+
+
+
+## [1.26.1](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/session-replay-browser@1.26.0...@amplitude/session-replay-browser@1.26.1) (2025-07-11)
+
+
+### Bug Fixes
+
+* **session replay:** better Angular zone detection ([#1205](https://github.com/amplitude/Amplitude-TypeScript/issues/1205)) ([7d88d29](https://github.com/amplitude/Amplitude-TypeScript/commit/7d88d29671810d294f809a87d642eec825959759))
+
+
+
+
+
 # [1.26.0](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/session-replay-browser@1.25.3...@amplitude/session-replay-browser@1.26.0) (2025-07-08)